# ai-control-core/acc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FluxForge ACC — AI Control Core",
    "description": "API for reading project files, searching code, and applying patches to FluxForge Studio. Used by ChatGPT Custom GPT as a failover when Claude Code is unavailable.",
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://YOUR_NGROK_URL.ngrok-free.app",
      "description": "ACC server via ngrok tunnel"
    }
  ],
  "paths": {
...
    "/gpt/files/read": {
      "get": {
        "operationId": "readFile",
        "summary": "Read a file from the project",
        "description": "Read the contents of a file. Path must be relative to repo root (e.g. 'crates/rf-engine/src/ffi.rs'). For large files, use line_start and line_end to read a specific range.",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Relative path from repo root (e.g. 'flutter_ui/lib/providers/mixer_provider.dart')"
          },
          {
            "name": "line_start",
            "in": "query",
            "required": false,
            "schema": {"type": "integer"},
            "description": "First line to read (1-indexed)"
          },
          {
            "name": "line_end",
            "in": "query",
            "required": false,
            "schema": {"type": "integer"},
            "description": "Last line to read (inclusive)"
          }
        ],
        "responses": {
          "200": {
            "description": "File content with line numbers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "path": {"type": "string"},
                    "total_lines": {"type": "integer"},
                    "showing_lines": {"type": "array", "items": {"type": "integer"}},
                    "content": {"type": "string"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gpt/files/tree": {
      "get": {
        "operationId": "listDirectory",
        "summary": "List directory contents as a tree",
        "description": "List files and subdirectories. Use to explore project structure before reading specific files.",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {"type": "string", "default": "."},
            "description": "Relative directory path (default: repo root)"
          },
          {
            "name": "depth",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 2, "minimum": 1, "maximum": 4},
            "description": "How deep to recurse (1-4)"
          }
        ],
        "responses": {
          "200": {
            "description": "Directory listing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "root": {"type": "string"},
                    "count": {"type": "integer"},
                    "truncated": {"type": "boolean"},
                    "entries": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "path": {"type": "string"},
                          "type": {"type": "string", "enum": ["file", "dir"]},
                          "size": {"type": "integer"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gpt/files/search": {
      "get": {
        "operationId": "searchCode",
        "summary": "Search code by text (grep)",
        "description": "Search for text in project files. Returns matching lines with context. Use glob to filter file types (e.g. '*.rs' for Rust, '*.dart' for Dart).",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "minLength": 2},
            "description": "Search text (min 2 chars)"
          },
          {
            "name": "glob",
            "in": "query",
            "required": false,
            "schema": {"type": "string"},
            "description": "File pattern filter (e.g. '*.rs', '*.dart', '*.toml')"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {"type": "string"},
            "description": "Search within this subdirectory (e.g. 'crates/' or 'flutter_ui/lib/')"
          },
          {
            "name": "context",
            "in": "query",
            "required": false,
            "schema": {"type": "integer", "default": 2, "minimum": 0, "maximum": 5},
            "description": "Lines of context around each match"
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {"type": "boolean"},
                    "query": {"type": "string"},
                    "match_count": {"type": "integer"},
                    "files_matched": {"type": "integer"},
                    "files": {"type": "array", "items": {"type": "string"}},
                    "matches": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "file": {"type": "string"},
                          "line": {"type": "integer"},
                          "content": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-acc-key"
      }
    }
  },
  "security": [
    {"apiKey": []}
  ]
}